+++ lib/fire_widgets.dart
@@ -1,2 +1,3 @@
 export 'src/expandable_bottom_sheet_page.dart';
 export 'src/theme_colors_viewer_page.dart';
+export 'src/expandable_card.dart';